- Päivittää profiilissa koiralistan
