+++ app/profileScreen.tsx
@@ -6,14 +6,34 @@ import { HStack } from '@/components/ui/hstack';
 import { Text } from '@/components/ui/text';
 import { VStack } from '@/components/ui/vstack';
 import { useAuth } from '@/src/context/AuthContext';
-import React from 'react';
+import { getMyDogs } from '@/src/services/dogService';
+import { useFocusEffect } from 'expo-router';
+import React, { useCallback, useState } from 'react';
 import { View } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
 
 export default function profileScreen() {
     const { userProfile } = useAuth()
+    const [dog, setDogs] = useState<any[]>([])
+    const [status, setStatus] = useState<string | null>(null)
     
+        const loadDogs = async () => {
+            setStatus('Ladataan koiria...')
+            try {
+                const list = await getMyDogs()
+                setDogs(list)
+                setStatus(null)
+            } catch (e: any) {
+                console.error(e)
+                setStatus(`Koirien lataus epäonnistui: ${e?.message ?? e}`)
+            }
+        }
 
+          useFocusEffect(
+    useCallback(() => {
+      loadDogs();
+    }, [])
+  );
     return (
         <SafeAreaView className="flex-1 px-4 pt-6">
             <View className="m-6 p-5 bg-white rounded-lg">
@@ -32,7 +52,7 @@ export default function profileScreen() {
             </View>
             <View className="mt-4">
                 <Text className="text-2xl font-bold text-white">My Dogs</Text>
-                <DogCard name={''} image={undefined} />
+                <DogCard dog={dog} name={''} image={undefined} />
             </View>
             <View className="mt-4">
                 <Text className="text-2xl font-bold text-white">My Teams</Text>
