+++ components/dogCard.tsx
@@ -1,44 +1,20 @@
 import { Card } from '@/components/ui/card';
 import { Heading } from '@/components/ui/heading';
-import { getMyDogs } from '@/src/services/dogService';
 import MaterialIcons from '@expo/vector-icons/MaterialIcons';
 import { router } from 'expo-router';
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { Image, ScrollView, TouchableOpacity, View } from 'react-native';
 
-export default function DogCard({ name, image }: { name: string; image: any }) {
-
-    const [status, setStatus] = useState<string | null>(null)
-    const [dogs, setDogs] = useState<any[]>([])
-
-    useEffect(() => {
-        loadDogs()
-    }, [])
-
-    const loadDogs = async () => {
-        setStatus('Ladataan koiria...')
-        try {
-            const list = await getMyDogs()
-            setDogs(list)
-            setStatus(null)
-        } catch (e: any) {
-            console.error(e)
-            setStatus(`Koirien lataus epäonnistui: ${e?.message ?? e}`)
-        }
-    }
-
-    const handleDogInfo = async () => {
-        console.log(dogs.map(dog => ({ name: dog.name, imageUrl: dog.imageUrl, id: dog.id, breed: dog.breed, age: dog.age, description: dog.description, gender: dog.gender, size: dog.size })))
-    }
+export default function DogCard({dog, name, image }: { dog: any[]; name: string; image: any }) {
 
     return (
         <View>
             <ScrollView horizontal showsHorizontalScrollIndicator={false} className="mt-4">
-                {dogs.map(dog => (
+                {dog.map(dog => (
                     <Card key={dog.id} size="md" variant="elevated" className="m-3 bg-gray-600 p-2 rounded-lg">
                         <TouchableOpacity
                             onPress={() =>
-                                router.push({
+                                router.replace({
                                     pathname: '/dogProfileScreen',
                                     params: {
                                         id: dog.id,
